pyproject.toml
[tool.poetry]
name = "Boilerplate"
version = "0.1.0"
description = "Thisi is a boilerplate for John Proud"
authors = ["ion <mindruyon@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
django-rest-framework = "^0.1.0"
django-parler = "^2.2"
soft-delete = "^0.2.2"
djangorestframework-simplejwt = "4.1.0"
django-cleanup = "^5.2.0"
django-environ = "^0.7.0"
django-celery-beat = "^2.2.1"
django-parler-rest = "^2.1"
pendulum = "^2.1.2"
djangorestframework = "^3.12.4"
djangorestframework-filters = "1.0.0.dev0"
drf-yasg = "^1.20.0"
geoip2 = "^4.2.0"
gunicorn = "^20.1.0"
Pillow = "^8.3.2"
django-cors-headers = "^3.8.0"
djet = "^0.3.0"
PyJWT = "1.7.1"
twilio = "^6.63.2"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
